pyproject.toml
[tool.poetry]
name = "exercism-python"
version = "0.1.0"
description = ""
authors = ["sprietl"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
docstring-parser = "^0.7.3"
comment-parser = "^1.2.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pylint = "^2.8.2"
pycodestyle = "^2.7.0"
autopep8 = "^1.5.7"
rope = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
